Cleaned up old Classes
Some old versions of my FXML files were still in the the program files, so I removed them. Also changed AbilityRaceBonusMenu to account for charisma being selected. Originally charisma wasn't an option, but I reused the scene for human variant where charisma is an option. I had to change the switch statement to account for this.